reminder/requirements.txt
amqp==2.5.1
arrow==0.15.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-celery-beat==1.5.0
django-filters==0.2.1
django-oauth-toolkit==1.2.0
django-timezone-field==3.0
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
kombu==4.6.4
more-itertools==7.2.0
oauthlib==3.1.0
optionaldict==0.1.1
psycopg2==2.8.3
pycryptodome==3.9.0
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.2
rachel==0.1.0
redis==2.10.0
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
sxtwl==1.0.7
urllib3==1.25.6
vine==1.3.0
wechatpy==1.8.3
Werkzeug==0.16.0
xmltodict==0.12.0
zipp==0.6.0
gunicorn==19.9.0
django-extensions==2.2.3


